Feat: 스팟 카테고리 등록 / 유효성 검사 로직 등록
- 스팟 카테고리 유효성 검사 로직 추가
- 스팟 카테고리 스팟에 등록 추가
- 스팟 관련 기타 서비스 함수 다른 파일에 분리리
